Fix latent type-inference bug in split heatpipesections clone
jsonb_build_object treats its value args as "any", so the bare $2 parameter
for 'lineid' left PostgreSQL unable to infer the parameter type
("could not determine data type of parameter $2"). This would have failed on
the very first real line split. Cast to $2::int.

Surfaced by the dry-run preview running the full split against the live DB:
line 375250 now returns review={dampers:1} and rolls back cleanly.

+++ itwin-api/database/topology.py
@@ -326,7 +326,9 @@ async def _split_line_body(conn, line_id: int, lng: float, lat: float) -> dict:
                   NULL::heatpipesections,
                   to_jsonb(h) || jsonb_build_object(
                     'id', nextval('heatpipesections_id_seq'),
-                    'lineid', $2,
+                    -- явный ::int: в jsonb_build_object аргумент имеет тип "any",
+                    -- и без каста PostgreSQL не может вывести тип параметра
+                    'lineid', $2::int,
                     'pipesectlength', ST_Length((SELECT shape FROM linesobj WHERE id = $2))
                   )
                 )).*
